Extract the compiler into a separate file so it can be included with no runtime dependency or pollution
