+++ nodent.js
@@ -7,11 +7,8 @@
  * AST transforms and node loader extension
  */
 var stdJSLoader ;
-var smCache = {} ;
 var fs = require('fs') ;
-var outputCode = require('./lib/output') ;
-var parser = require('./lib/parser') ;
-var treeSurgeon = require('./lib/arboriculture') ;
+var NodentCompiler = require('./compiler') ;
 
 // Config options used to control the run-time behaviour of the compiler
 var config = {
@@ -42,25 +39,6 @@ var config = {
 //		named by the 'use nodent-OPTION', as set by the setCompileOptions (or setDefaultCompileOptions)
 //		set within the directive as a JSON-encoded extension
 
-var initialCodeGenOpts = {
-    noRuntime:false,
-    lazyThenables:false,
-    es6target:false,
-    noUseDirective:false,
-	wrapAwait:null,
-	mapStartLine:0,
-	sourcemap:true,
-    engine:false,
-	parser:{sourceType:'script'},
-	$return:"$return",
-	$error:"$error",
-	$arguments:"$args",
-	$asyncspawn:"$asyncspawn",
-	$asyncbind:"$asyncbind",
-	generatedSymbolPrefix:"$",
-	$makeThenable:'$makeThenable'
-};
-
 function copyObj(a){
 	var o = {} ;
 	a.forEach(function(b){
@@ -71,7 +49,7 @@ function copyObj(a){
 	return o ;
 };
 
-var defaultCodeGenOpts = Object.create(initialCodeGenOpts, {es7:{value:true,writable:true,enumerable:true}}) ;
+var defaultCodeGenOpts = Object.create(NodentCompiler.initialCodeGenOpts, {es7:{value:true,writable:true,enumerable:true}}) ;
 var optionSets = {
 	default:defaultCodeGenOpts,
 	es7:Object.create(defaultCodeGenOpts),
@@ -100,8 +78,6 @@ var runtimes = require('nodent-runtime') ;
 var $asyncbind = runtimes.$asyncbind ;
 var $asyncspawn = runtimes.$asyncspawn ;
 var Thenable = $asyncbind.Thenable ;
-
-function noLogger(){}
 
 function isDirective(node){
     return node.type === 'ExpressionStatement' &&
@@ -186,17 +162,6 @@ function stripBOM(content) {
 	return content;
 }
 
-function btoa(str) {
-	var buffer ;
-	if (str instanceof Buffer) {
-		buffer = str;
-	} else {
-		buffer = new Buffer(str.toString(), 'binary');
-	}
-
-	return buffer.toString('base64');
-}
-
 function compileNodentedFile(nodent,log) {
     log = log || nodent.log ;
 	return function(mod, filename, parseOpts) {
@@ -266,107 +231,6 @@ function asyncify(promiseProvider) {
 	}
 };
 
-function prettyPrint(pr,opts) {
-	var map ;
-	var filepath = pr.filename.split("/") ;
-	var filename = filepath.pop() ;
-
-	var out = outputCode(pr.ast,(opts && opts.sourcemap)?{map:{
-		startLine: opts.mapStartLine || 0,
-		file: filename+"(original)",
-		sourceMapRoot: filepath.join("/"),
-		sourceContent: pr.origCode
-	}}:null, pr.origCode) ;
-
-	if (opts && opts.sourcemap){
-	    try {
-	        var mapUrl = "" ;
-	        var jsmap = out.map.toJSON();
-	        if (jsmap) {
-	            // require an expression to defeat browserify
-	            var SourceMapConsumer = require('source-map').SourceMapConsumer;
-	            pr.sourcemap = jsmap ;
-	            smCache[pr.filename] = {map:jsmap,smc:new SourceMapConsumer(jsmap)} ;
-	            mapUrl = "\n\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,"
-	                +btoa(JSON.stringify(jsmap))+"\n" ;
-	        }
-	        pr.code = out.code+mapUrl ;
-	    } catch (ex) {
-	        pr.code = out ;
-	    }
-	} else {
-        pr.code = out ;
-	}
-	return pr ;
-}
-
-function parseCode(code,origFilename,__sourceMapping,opts) {
-	if (typeof __sourceMapping==="object" && opts===undefined)
-		opts = __sourceMapping ;
-
-	var r = { origCode:code.toString(), filename:origFilename } ;
-	try {
-		r.ast = parser.parse(r.origCode, opts && opts.parser) ;
-		if (opts.babelTree) {
-			parser.treeWalker(r.ast,function(node,descend,path){
-				if (node.type==='Literal')
-					path[0].replace(treeSurgeon.babelLiteralNode(node.value)) ;
-				else if (node.type==='Property') {
-				    // Class/ObjectProperty in babel6
-				    if (path[0].parent.type==='ClassBody'){
-				        // There's no easy mapping here as it appears to be borderline in the specification?
-				        // It's definitely a kind of ClassProperty tho....
-                        node.type = 'ClassProperty' ;
-				    } else {
-				        node.type = 'ObjectProperty' ;
-				    }
-				}
-				descend() ;
-			}) ;
-		}
-		return r ;
-	} catch (ex) {
-		if (ex instanceof SyntaxError) {
-			var l = r.origCode.substr(ex.pos-ex.loc.column) ;
-			l = l.split("\n")[0] ;
-			ex.message += " "+origFilename+" (nodent)\n"+l+"\n"+l.replace(/[\S ]/g,"-").substring(0,ex.loc.column)+"^" ;
-			ex.stack = "" ;
-		}
-		throw ex ;
-	}
-}
-
-/* NodentCompiler prototypes, that refer to 'this' */
-function requireCover(cover,opts) {
-	opts = opts || {} ;
-	var key = cover+'|'+Object.keys(opts).sort().reduce(function(a,k){ return a+k+JSON.stringify(opts[k])},"") ;
-	if (!this.covers[key]) {
-		if (cover.indexOf("/")>=0)
-			this.covers[key] = require(cover) ;
-		else
-			this.covers[key] = require(__dirname+"/covers/"+cover);
-	}
-	return this.covers[key](this,opts) ;
-}
-
-function compile(code,origFilename,__sourceMapping,opts) {
-	if (typeof __sourceMapping==="object" && opts===undefined)
-		opts = __sourceMapping ;
-
-	opts = opts || {} ;
-
-	// Fill in any default codeGen options
-	for (var k in defaultCodeGenOpts) {
-		if (!(k in opts))
-			opts[k] = defaultCodeGenOpts[k] ;
-	}
-
-	var pr = this.parse(code,origFilename,null,opts);
-	this.asynchronize(pr,null,opts,this.log || noLogger) ;
-	this.prettyPrint(pr,opts) ;
-	return pr ;
-}
-
 function generateRequestHandler(path, matchRegex, options) {
 	var cache = {} ;
 	var compiler = this ;
@@ -377,7 +241,7 @@ function generateRequestHandler(path, matchRegex, options) {
 		options = {compiler:{}} ;
 	else if (!options.compiler)
 		options.compiler = {} ;
-	var compilerOptions = copyObj([initialCodeGenOpts,options.compiler]) ;
+	var compilerOptions = copyObj([NodentCompiler.initialCodeGenOpts,options.compiler]) ;
 
 	return function (req, res, next) {
 		if (cache[req.url]) {
@@ -445,53 +309,35 @@ function generateRequestHandler(path, matchRegex, options) {
 	};
 };
 
-function NodentCompiler(members) {
-	this.covers = {} ;
-	this._ident = NodentCompiler.prototype.version+"_"+Math.random() ;
-	this.setOptions(members) ;
+function requireCover(cover,opts) {
+    opts = opts || {} ;
+    var key = cover+'|'+Object.keys(opts).sort().reduce(function(a,k){ return a+k+JSON.stringify(opts[k])},"") ;
+    if (!this.covers[key]) {
+        if (cover.indexOf("/")>=0)
+            this.covers[key] = require(cover) ;
+        else
+            this.covers[key] = require(__dirname+"/covers/"+cover);
+    }
+    return this.covers[key](this,opts) ;
 }
 
-NodentCompiler.prototype.setOptions = function(members){
-	this.log = members.log===false?noLogger:members.log||this.log;
-	this.options = copyObj([this.options,members]) ;
-	delete this.options.log ;
-	return this ;
-};
-
-$asyncbind.call($asyncbind) ;
-
-NodentCompiler.prototype.version =  require("./package.json").version ;
 NodentCompiler.prototype.Thenable = Thenable ;
 NodentCompiler.prototype.EagerThenable = $asyncbind.EagerThenableFactory ;
-NodentCompiler.prototype.isThenable = function(x) { return x && x instanceof Object && typeof x.then==="function"} ;
 NodentCompiler.prototype.asyncify =  asyncify ;
 NodentCompiler.prototype.require =  requireCover ;
 NodentCompiler.prototype.generateRequestHandler = generateRequestHandler ;
 // Exported so they can be transported to a client
 NodentCompiler.prototype.$asyncspawn =  $asyncspawn ;
 NodentCompiler.prototype.$asyncbind =  $asyncbind ;
-// Exported ; but not to be used lightly!
-NodentCompiler.prototype.parse = parseCode ;
-NodentCompiler.prototype.compile =  compile ;
-NodentCompiler.prototype.asynchronize =  treeSurgeon.asynchronize ;
-NodentCompiler.prototype.prettyPrint =  prettyPrint ;
 NodentCompiler.prototype.parseCompilerOptions =  parseCompilerOptions ;
-NodentCompiler.prototype.getDefaultCompileOptions = undefined ;
 
-Object.defineProperty(NodentCompiler.prototype,"Promise",{
-	get:function (){
-		initOpts.log("Warning: nodent.Promise is deprecated. Use nodent.Thenable instead");
-		return Thenable;
-	},
-	enumerable:false,
-	configurable:false
-}) ;
+$asyncbind.call($asyncbind) ;
 
 function prepareMappedStackTrace(error, stack) {
 	function mappedTrace(frame) {
 		var source = frame.getFileName();
-		if (source && smCache[source]) {
-			var position = smCache[source].smc.originalPositionFor({
+		if (source && NodentCompiler.prototype.smCache[source]) {
+			var position = NodentCompiler.prototype.smCache[source].smc.originalPositionFor({
 				line: frame.getLineNumber(),
 				column: frame.getColumnNumber()
 			});
