+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "nodent",
-  "version": "3.0.20",
+  "version": "3.1.0",
   "description": "NoDent - Asynchronous Javascript language extensions",
   "main": "nodent.js",
   "scripts": {
